backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart==0.0.6
supabase>=2.27.0
pillow>=11.0.0
numpy==1.26.3
httpx==0.26.0
pydantic>=2.10.0
python-dotenv==1.0.0
onnxruntime>=1.20.0
slowapi>=0.1.0